Make the skin deployment deterministic
Before, the code implicitly relied on 'archive', which defaults to
'true' and implies:

  Mirrors the rsync archive flag, enables recursive, links, perms,
  times, owner, group flags and -D.

This made the deployment nondeterministic in environments which rely on
git, as that doesn't track times: results would be bound to the fact
*when* a git checkout happened.

This change ignores times, via 'archive' set to 'false'. Instead, it
relies on 'checksum', which is reliable. Besides, 'recursive' is set to
'true' explicitly, to mirror the former behaviour.